aarch64: NMI: support NMI for P660
The NMI could deal with the process when the irq is disabled,
but in aarch64, the irq could not be used for NMI becasue it couldn't
work when irq mask is set, so we need to use fiq as the nmi for system.

The normal irq could be set to fiq and route to EL3(trust firmware layer),
The EL3 will jump to the NMI hander which was registered to the EL3 and deal
with the handler, then use smc() return to the EL3 and restore the interrupt
context and return to normal work.

This patch was mainly devide to two parts, 1) is used for connect to interactive
with the EL3. 2) is used to add the NMI watchdog structure to enable NMI for OS.

The NMI has 3 parameters: 1) timeout, used to set the timer period for watchdog.
2) state, used to show and set the state for NMI percpu. 3) nmi_handler, the main
handler for NMI system.

Signed-off-by: Ding Tianhong <dingtianhong@huawei.com>