hip05: gmac: fix the panic when reset the gmac
When reset the gmac, the system may panic:

[260142.367696] hip05-mac c2080000.ethernet gmac0: inconsistent tx_skb
[260142.420013] hip05-mac c2080000.ethernet gmac0: inconsistent tx_skb
[260142.480708] hip05-mac c2080000.ethernet gmac0: inconsistent tx_skb
[260142.656855] hip05-mac c2080000.ethernet gmac0: inconsistent tx_skb
[260145.875413] BUG: failure at lib/dynamic_queue_limits.c:26/dql_completed()!

The reason is when xmit a skb, the driver will xmit first and then report xmit len
to the stack, so when reclaim the xmit queue, the driver will report the completed skb
but which is not report to xmit yet, so panic happened.

Fix this by reporting the xmit skb before xmit happened.

Signed-off-by: Ding Tianhong <dingtianhong@huawei.com>