3.19/4.1 Fixed the mbigen bug when device driver try to free a interrupt,connect to mbigen using free_irq()
Fixed the mbigen bug when device driver try to free a interrupt
 connect to mbigen using free_irq()

Assume a device with two interrupts A ,B connects to Mbigen.
Because of this bug, When device driver try to free A interrupt
using free_irq(A), the device id is corrupted .
So the interrupt B can't trigger neither.

Signed-off-by: Ma Jun <majun258@huawei.com>
Signed-off-by: Ding Tianhong <dingtianhong@huawei.com>