gic-v3: use new method to determine which node a SPI should be binding to
The latest numa have not binding to hisi hardware yet. But the SPI number
is associated with DIE-ID. That is, SPI interrupt number = die-id * 128
+ xxx, xxx < 128. Hardware designer suggested that: binding all SPIs to
their local DIE, because of GIC hardware BUG. We extract DIE-ID from a
SPI, and use it to match a cpu, then use cpu_to_node to get nid.

Limit:
All CPUs on the same DIE should also on the same numa node.

Signed-off-by: Zhen Lei <thunder.leizhen@huawei.com>
Signed-off-by: Hanjun Guo <guohanjun@huawei.com>
Signed-off-by: Ding Tianhong <dingtianhong@huawei.com>